backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
google-generativeai==0.7.0
faiss-cpu==1.12.0
sentence-transformers>=2.2.2
transformers>=4.30.0

requests==2.31.0
pypdf==6.0.0
